+++ client/src/commonMain/kotlin/eu/vendeli/rethis/command/serde/HVals.kt
@@ -4,6 +4,7 @@ import eu.vendeli.rethis.ReThis
 import eu.vendeli.rethis.command.hash.hVals
 import eu.vendeli.rethis.shared.types.DataProcessingException
 import eu.vendeli.rethis.types.interfaces.SerializationFormat
+import eu.vendeli.rethis.utils.isInTx
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.serializer
 
@@ -16,6 +17,10 @@ suspend fun <T : Any> ReThis.hVals(
     serializer: KSerializer<T>,
     format: SerializationFormat = cfg.serializationFormat,
 ): List<T> {
+    if (isInTx()) {
+        logger.warn("Be aware that in transaction commands return `QUEUED`" +
+            " which is for type safety substituted with default value, so serde operations will fail")
+    }
     val raw: List<String> = hVals(key)
     return raw.map { string ->
         try {