+++ client/src/commonMain/kotlin/eu/vendeli/rethis/command/serde/Mset.kt
@@ -4,6 +4,7 @@ import eu.vendeli.rethis.ReThis
 import eu.vendeli.rethis.command.string.mSet
 import eu.vendeli.rethis.shared.request.string.KeyValue
 import eu.vendeli.rethis.types.interfaces.SerializationFormat
+import eu.vendeli.rethis.utils.isInTx
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.serializer
 
@@ -16,6 +17,10 @@ suspend fun <T : Any> ReThis.mSet(
     serializer: KSerializer<T>,
     format: SerializationFormat = cfg.serializationFormat,
 ): Boolean {
+    if (isInTx()) {
+        logger.warn("Be aware that in transaction commands return `QUEUED`" +
+            " which is for type safety substituted with default value, so serde operations will fail")
+    }
     val serializedPairs = kvPair.map { (k, v) ->
         KeyValue(k, format.serialize(serializer, v))
     }.toTypedArray()