+++ client/src/commonMain/kotlin/eu/vendeli/rethis/command/serde/Set.kt
@@ -4,6 +4,7 @@ import eu.vendeli.rethis.ReThis
 import eu.vendeli.rethis.command.string.set
 import eu.vendeli.rethis.shared.request.string.SetOption
 import eu.vendeli.rethis.types.interfaces.SerializationFormat
+import eu.vendeli.rethis.utils.isInTx
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.serializer
 
@@ -22,6 +23,10 @@ suspend fun <T> ReThis.`set`(
     vararg options: SetOption,
     serializationFormat: SerializationFormat = cfg.serializationFormat,
 ): String? {
+    if (isInTx()) {
+        logger.warn("Be aware that in transaction commands return `QUEUED`" +
+            " which is for type safety substituted with default value, so serde operations will fail")
+    }
     val value = serializationFormat.serialize(serializer, value)
     return set(key, value, *options)
 }