+++ client/src/jvmTest/kotlin/eu/vendeli/rethis/utils/RTypeResponseTest.kt
@@ -1,28 +1,28 @@
-//package eu.vendeli.rethis.utils
+// package eu.vendeli.rethis.utils
 //
-//import com.ionspin.kotlin.bignum.integer.toBigInteger
-//import eu.vendeli.rethis.ReThisTestCtx
-//import eu.vendeli.rethis.shared.types.BigNumber
-//import eu.vendeli.rethis.shared.types.Bool
-//import eu.vendeli.rethis.shared.types.BulkString
-//import eu.vendeli.rethis.shared.types.F64
-//import eu.vendeli.rethis.shared.types.Int64
-//import eu.vendeli.rethis.shared.types.PlainString
-//import eu.vendeli.rethis.shared.types.Push
-//import eu.vendeli.rethis.shared.types.RArray
-//import eu.vendeli.rethis.shared.types.RMap
-//import eu.vendeli.rethis.shared.types.RSet
-//import eu.vendeli.rethis.shared.types.RType
-//import eu.vendeli.rethis.shared.types.VerbatimString
-//import eu.vendeli.rethis.utils.response.parseResponse
-//import eu.vendeli.rethis.utils.response.readResponseWrapped
-//import eu.vendeli.rethis.utils.safeCast
-//import io.kotest.matchers.shouldBe
-//import io.ktor.utils.io.*
-//import io.ktor.utils.io.core.*
-//import kotlinx.io.Buffer
+// import com.ionspin.kotlin.bignum.integer.toBigInteger
+// import eu.vendeli.rethis.ReThisTestCtx
+// import eu.vendeli.rethis.shared.types.BigNumber
+// import eu.vendeli.rethis.shared.types.Bool
+// import eu.vendeli.rethis.shared.types.BulkString
+// import eu.vendeli.rethis.shared.types.F64
+// import eu.vendeli.rethis.shared.types.Int64
+// import eu.vendeli.rethis.shared.types.PlainString
+// import eu.vendeli.rethis.shared.types.Push
+// import eu.vendeli.rethis.shared.types.RArray
+// import eu.vendeli.rethis.shared.types.RMap
+// import eu.vendeli.rethis.shared.types.RSet
+// import eu.vendeli.rethis.shared.types.RType
+// import eu.vendeli.rethis.shared.types.VerbatimString
+// import eu.vendeli.rethis.utils.response.parseResponse
+// import eu.vendeli.rethis.utils.response.readResponseWrapped
+// import eu.vendeli.rethis.utils.safeCast
+// import io.kotest.matchers.shouldBe
+// import io.ktor.utils.io.*
+// import io.ktor.utils.io.core.*
+// import kotlinx.io.Buffer
 //
-//class RTypeResponseTest : ReThisTestCtx() {
+// class RTypeResponseTest : ReThisTestCtx() {
 //    @Test
 //    suspend fun `test readRedisMessage with simple string`() {
 //        val channel = ByteReadChannel {
@@ -207,4 +207,4 @@
 //
 //        return ByteReadChannel(buff)
 //    }
-//}
+// }