+++ client/src/jvmTest/kotlin/eu/vendeli/rethis/utils/ResponseUtilsTest.kt
@@ -1,22 +1,22 @@
-//package eu.vendeli.rethis.utils
+// package eu.vendeli.rethis.utils
 //
-//import com.ionspin.kotlin.bignum.integer.BigInteger
-//import com.ionspin.kotlin.bignum.integer.toBigInteger
-//import eu.vendeli.rethis.ReThisTestCtx
-//import eu.vendeli.rethis.shared.types.ReThisException
-//import eu.vendeli.rethis.utils.response.parseResponse
-//import eu.vendeli.rethis.utils.response.readListResponseTyped
-//import eu.vendeli.rethis.utils.response.readMapResponseTyped
-//import eu.vendeli.rethis.utils.response.readSimpleResponseTyped
-//import io.kotest.assertions.throwables.shouldThrow
-//import io.kotest.matchers.nulls.shouldBeNull
-//import io.kotest.matchers.shouldBe
-//import io.ktor.util.reflect.*
-//import io.ktor.utils.io.*
-//import io.ktor.utils.io.core.*
-//import kotlinx.io.Buffer
+// import com.ionspin.kotlin.bignum.integer.BigInteger
+// import com.ionspin.kotlin.bignum.integer.toBigInteger
+// import eu.vendeli.rethis.ReThisTestCtx
+// import eu.vendeli.rethis.shared.types.ReThisException
+// import eu.vendeli.rethis.utils.response.parseResponse
+// import eu.vendeli.rethis.utils.response.readListResponseTyped
+// import eu.vendeli.rethis.utils.response.readMapResponseTyped
+// import eu.vendeli.rethis.utils.response.readSimpleResponseTyped
+// import io.kotest.assertions.throwables.shouldThrow
+// import io.kotest.matchers.nulls.shouldBeNull
+// import io.kotest.matchers.shouldBe
+// import io.ktor.util.reflect.*
+// import io.ktor.utils.io.*
+// import io.ktor.utils.io.core.*
+// import kotlinx.io.Buffer
 //
-//class ResponseUtilsTest : ReThisTestCtx() {
+// class ResponseUtilsTest : ReThisTestCtx() {
 //    @Test
 //    suspend fun `test processRedisSimpleResponse with simple string`() {
 //        val channel = ByteReadChannel {
@@ -182,4 +182,4 @@
 //
 //        return ByteReadChannel(buff)
 //    }
-//}
+// }